Added and moved the directory structure
