+++ azure-pipelines.yml
@@ -1,4 +1,4 @@
-name: 'ISO3166UtilLib build'
+name: 'TimeThis build'
 
 trigger:
   batch: true
@@ -100,14 +100,14 @@ steps:
 - task: GitHubRelease@1
   inputs:
     gitHubConnection: 'github-packages-sqs'
-    repositoryName: 'siddiqsoft/ISO3166UtilLib'
+    repositoryName: 'siddiqsoft/TimeThis'
     action: 'create'
     target: '$(Build.SourceVersion)'
     tagSource: 'userSpecifiedTag'
     tag: '$(build.buildNumber)'
     title: 'v$(build.buildNumber)'
     releaseNotesSource: 'inline'
-    releaseNotesInline: 'Documentation https://siddiqsoft.github.io/ISO3166UtilLib'
+    releaseNotesInline: 'Documentation https://siddiqsoft.github.io/TimeThis'
     isPreRelease: true
     changeLogCompareToRelease: 'lastFullRelease'
     changeLogType: 'commitBased'
