+++ testEnvironments.json
@@ -2,21 +2,9 @@
   "version": "1",
   "environments": [
     {
-      "name": "lba1.local",
+      "name": "nuc2.local",
       "type": "ssh",
-      "remoteUri": "ssh://kei@lba1.local:22"
+      "remoteUri": "ssh://maas@siddiq.org@nuc2.local:22"
     }
-    // See https://aka.ms/remotetesting for more details
-    // about how to configure remote environments.
-    //{
-    //  "name": "WSL Ubuntu",
-    //  "type": "wsl",
-    //  "wslDistribution": "Ubuntu"
-    //},
-    //{
-    //  "name": "Docker dotnet/sdk",
-    //  "type": "docker",
-    //  "dockerImage": "mcr.microsoft.com/dotnet/sdk"
-    //}
   ]
 }
