+++ tests/CMakeLists.txt
@@ -10,17 +10,18 @@ include(FetchContent)
 FetchContent_Declare(googletest
 					GIT_REPOSITORY https://github.com/google/googletest.git
 					GIT_TAG	v1.15.2
+					FIND_PACKAGE_ARGS NAMES GTest
 )
 FetchContent_MakeAvailable(googletest)
 
+# Declare our test source files.
 file(GLOB testSRC "./test.cpp")
 
-add_executable(ISO3166UtilLib_test ${testSRC})
+add_executable(TimeThis_test ${testSRC})
 
-target_link_libraries(ISO3166UtilLib_test
- PRIVATE GTest::gtest_main)
+target_link_libraries(TimeThis_test PRIVATE GTest::gtest_main)
 
 include(GoogleTest)
-gtest_discover_tests(ISO3166UtilLib_test
-                     XML_OUTPUT_DIR "${PROJECT_SOURCE_DIR}/TestResults"
+gtest_discover_tests(TimeThis_test
+                     XML_OUTPUT_DIR "${PROJECT_SOURCE_DIR}/tests/results"
 )
